MAJOR CHANGE:
changed analysis routines to return dictionaries instead of arrays.

Key: string to be appended to the save label
Value: array to be saved.

for e.g., the statement: "{'':auc_data*bbox_mask}"

returns an auc_data and the scan label to save the data with is the same as label given.