feature: allow direct attribute access to Experiment.patients and Experiment.labels
this has the tremendously pleasant effect of allowing tab completion on patient names as well as scan labels. Note that this will require the use of python syntax-compliant naming of patients and labels (no spaces or minus signs). This is a small price to pay. Alternatively, the usual string-based key access still works.